moved a function from debug to utils_general
