+++ Makefile
@@ -25,15 +25,15 @@ OBJ_PATH = ./obj/
 UTILS_PATH = ./src/
 
 # SOURCES
-SRC =	$(SRC_PATH)main.c			$(SRC_PATH)lexer.c\
-	$(SRC_PATH)utils_error.c		$(SRC_PATH)utils_lexer.c\
-	$(SRC_PATH)signals.c			$(SRC_PATH)utils_free.c\
-	$(SRC_PATH)utils_lexer2.c		$(SRC_PATH)parser.c\
-	$(SRC_PATH)utils_parser.c		$(SRC_PATH)executer.c\
-	$(SRC_PATH)builtins.c			$(SRC_PATH)utils_parser_types.c\
-	$(SRC_PATH)utils_parser_dquote.c 	$(SRC_PATH)export.c\
-	$(SRC_PATH)utils_parser_command.c $(SRC_PATH)utils_export.c\
-	$(SRC_PATH)unset.c\
+SRC =	$(SRC_PATH)main.c				$(SRC_PATH)lexer.c\
+	$(SRC_PATH)utils_error.c			$(SRC_PATH)utils_lexer.c\
+	$(SRC_PATH)signals.c				$(SRC_PATH)utils_free.c\
+	$(SRC_PATH)utils_lexer2.c			$(SRC_PATH)parser.c\
+	$(SRC_PATH)utils_parser.c			$(SRC_PATH)executer.c\
+	$(SRC_PATH)builtins.c				$(SRC_PATH)utils_parser_types.c\
+	$(SRC_PATH)utils_parser_dquote.c	$(SRC_PATH)export.c\
+	$(SRC_PATH)utils_parser_command.c	$(SRC_PATH)utils_export.c\
+	$(SRC_PATH)unset.c					$(SRC_PATH)utils_general.c\
 	$(SRC_PATH)debug.c
 
 # OBJECTS
